#48 implemented limitload integral
