+++ openmodelica_microgrid_gym/net/components.py
@@ -1,130 +1,10 @@
-from itertools import chain
-from typing import Optional, Dict
-
 import numpy as np
 
 from openmodelica_microgrid_gym.aux_ctl import DDS, DroopController, DroopParams, InverseDroopController, \
     InverseDroopParams, PLLParams, PLL
-from openmodelica_microgrid_gym.net.net import Network
+from openmodelica_microgrid_gym.aux_ctl.base import LimitLoadIntegral
+from openmodelica_microgrid_gym.net.base import Component
 from openmodelica_microgrid_gym.util import dq0_to_abc, inst_power, inst_reactive
-
-
-class Component:
-    def __init__(self, net: Network, id=None, in_vars=None, out_vars=None, out_calc=None):
-        """
-
-        :param net:
-        :param id:
-        :param in_vars:
-        :param out_vars:
-        :param out_calc: mapping from attr name to
-        """
-        self.net = net
-        self.id = id
-        for attr in chain.from_iterable((f.keys() for f in filter(None, (in_vars, out_vars)))):
-            if not hasattr(self, attr):
-                raise AttributeError(f'{self.__class__} no such attribute: {attr}')
-        self.in_vars = in_vars
-        self.in_idx = None  # type: Optional[dict]
-
-        self.out_calc = out_calc or {}
-        self.out_vars = out_vars
-        self.out_idx = None  # type: Optional[dict]
-
-    def reset(self):
-        pass
-
-    def params(self, actions):
-        """
-        Calculate additional environment parameters
-        :param actions:
-        :return: mapping
-        """
-        return {}
-
-    def get_in_vars(self):
-        """
-        list of input variable names of this component
-        """
-        if self.in_vars:
-            return [[self._prefix_var(val) for val in vals] for attr, vals in self.in_vars.items()]
-        return []
-
-    def get_out_vars(self, with_aug=False):
-        r = []
-        if self.out_vars:
-            r = [[self._prefix_var(val) for val in vals] for attr, vals in self.out_vars.items()]
-
-        if not with_aug:
-            return r
-        else:
-            return r + [[self._prefix_var([self.id, attr, str(i)]) for i in range(n)] for attr, n in
-                        self.out_calc.items()]
-
-    def fill_tmpl(self, state):
-        if self.out_idx is None:
-            raise ValueError('call set_tmplidx before fill_tmpl. the keys must be converted to indices for efficiency')
-        for attr, idxs in self.out_idx.items():
-            # set object variables to the respective state variables
-            if hasattr(self, attr):
-                setattr(self, attr, state[idxs])
-            else:
-                raise AttributeError(f'{self.__class__} has no such attribute: {attr}')
-
-    def set_outidx(self, keys):
-        # This pre-calculation is done mainly for performance reasons
-        keyidx = {v: i for i, v in enumerate(keys)}
-        self.out_idx = {}
-        try:
-            for var, keys in self.out_vars.items():
-                # lookup index in the whole state keys
-                self.out_idx[var] = [keyidx[self._prefix_var(key)] for key in keys]
-        except KeyError as e:
-            raise KeyError(f'the output variable {e!s} is not provided by your state keys')
-
-    def _prefix_var(self, strs):
-        if isinstance(strs, str):
-            strs = [strs]
-        if strs[0].startswith('.'):
-            # first string minus its prefix '.' and the remaining strs
-            strs[0] = strs[0][1:]
-            if self.id is not None:
-                # this is a complete identifier like 'lc1.inductor1.i' that should not be modified:
-                strs = [self.id] + strs
-        return '.'.join(strs)
-
-    def calculate(self) -> Dict[str, np.ndarray]:
-        """
-        will write internal variables it is called after all internal variables are set
-        The return value must be a dictionary whose keys match the keys of self.out_calc and whose values are of the length of outcalcs values
-
-        set(self.out_calc.keys()) == set(return)
-        all([len(v) == self.out_calc[k] for k,v in return.items()])
-        :return:
-        """
-        pass
-
-    def normalize(self, calc_data):
-        pass
-
-    def augment(self, state, normalize=True):
-        self.fill_tmpl(state)
-        calc_data = self.calculate()
-
-        if normalize:
-            self.normalize(calc_data)
-        attr = ''
-        try:
-            new_vals = []
-            for attr, n in self.out_calc.items():
-                for i in range(n):
-                    new_vals.append(calc_data[attr][i])
-            return np.hstack([getattr(self, attr) for attr in self.out_idx.keys()] + new_vals)
-        except KeyError as e:
-            raise ValueError(
-                f'{self.__class__} missing return key: {e!s}. did you forget to set it in the calculate method?')
-        except IndexError as e:
-            raise ValueError(f'{self.__class__}.calculate()[{attr}] has the wrong number of values')
 
 
 class Inverter(Component):
@@ -138,14 +18,26 @@ class Inverter(Component):
         self.v_DC = v_DC
         self.i_ref = i_ref
         super().__init__(**{'out_calc': dict(i_ref=3), **kwargs})
+        self.limit_load_integrals = [
+            LimitLoadIntegral(self.net.ts, self.net.freq_nom, i_nom=i_nom, i_lim=i_lim) for _ in
+            range(3)]
+
+    def reset(self):
+        [integ.reset() for integ in self.limit_load_integrals]
 
     def normalize(self, calc_data):
         self.i /= self.i_lim
         self.v /= self.v_lim
         calc_data['i_ref'] /= self.i_lim
 
+    def risk(self) -> float:
+        return max([integ.risk() for integ in self.limit_load_integrals])
+
     def params(self, actions):
         return {**super().params(actions), **{self._prefix_var(['.v_DC']): self.v_DC}}
+
+    def calculate(self):
+        [integ.step(i) for i, integ in zip(self.i, self.limit_load_integrals)]
 
 
 class SlaveInverter(Inverter):
@@ -169,6 +61,7 @@ class SlaveInverter(Inverter):
         self.pll.reset()
 
     def calculate(self):
+        super().calculate()
         _, _, phase = self.pll.step(self.v)
         return dict(i_ref=dq0_to_abc(self.i_ref, phase))
