+++ tests/test_runner.py
@@ -117,8 +117,8 @@ def test_simpleagent(env):
     runner.run(1)
 
     # env.history.df.to_hdf('tests/test_main3.hd5', 'hist')
-    df = env.history.df.head(50)
+    df = env.history.df.head(40)
     df = df.reindex(sorted(df.columns), axis=1)
-    df2 = pd.read_hdf('tests/test_main3.hd5', 'hist').head(50)  # noqa
+    df2 = pd.read_hdf('tests/test_main3.hd5', 'hist').head(40)  # noqa
     df2 = df2.reindex(sorted(df2.columns), axis=1)
-    assert df[out_params].to_numpy() == approx(df2[out_params].to_numpy(), 5e-3)
+    assert df[out_params].to_numpy() == approx(df2[out_params].to_numpy(), 5e-4)
